Moved sql templates to ./sql and reorganized trade-areas analisys
