+++ lib/node/index.js
@@ -6,7 +6,7 @@ var nodes = {
     PointInPolygon: require('./nodes/point-in-polygon'),
     PopulationInArea: require('./nodes/population-in-area'),
     Source: require('./nodes/source'),
-    TradeArea: require('./nodes/trade-area/trade-area')
+    TradeArea: require('./nodes/trade-area')
 };
 
 module.exports = nodes;
