+++ lib/node/nodes/sql/trade-area-dissolved.sql
