+++ lib/node/nodes/sql/trade-area.sql
